alteracao do nome da pasta com scripts
