+++ spotconnect/bash_scripts.py
