+++ spotconnect/data/key_pair_default_dir.txt
