+++ spotconnect/data/spot_instance_pricing.csv
