+++ spotconnect/s3_methods.py
@@ -12,51 +12,52 @@ MIT License 2020
 """
 
 import os
-from path import Path 
+from path import Path
 
 root = Path(os.path.dirname(os.path.abspath(__file__)))
 
-from spot_connect.sutils import chunks
+from spotconnect.sutils import chunks
 
 import boto3
-from Ipython.display import clear_output 
+from Ipython.display import clear_output
+
 
 def listS3Objects(bucket_name):
     s3 = boto3.resource('s3')
 
     bucket = s3.Bucket(bucket_name)
 
-    s3_files = [] 
-    i = 0 
+    s3_files = []
+    i = 0
     for f in bucket.objects.all():
         s3_files.append(f)
-        i+=1 
-        if i%100==0:
+        i += 1
+        if i % 100 == 0:
             print(f)
             print('object #:', i)
             clear_output(wait=True)
 
-    return s3_files 
+    return s3_files
 
-def deleteS3Objects(bucket_name : str, objects : list): 
+
+def deleteS3Objects(bucket_name: str, objects: list):
     '''
     Delete a list of s3 objects from a given s3 bucket
     '''
 
     assert type(objects[0]) == boto3.resources.factory.s3.ObjectSummary
-    
+
     s3 = boto3.resource('s3')
 
     bucket = s3.Bucket(bucket_name)
 
-
     file_chunks = [c for c in chunks(objects, 1000)]
-    
+
     responses = []
-    
-    for files in file_chunks: 
-        ftod = [{'Key':c.key} for c in files]
+
+    for files in file_chunks:
+        ftod = [{'Key': c.key} for c in files]
         response = bucket.delete_objects(
             Delete={'Objects': ftod}
-        ) 
+        )
         responses.append(response)
